feat: 댓글 에러 핸들링 및 데이터 검증 추가

+++ Personal Mission/mission4/using react nodejs blog/backend/service/commentService.js
@@ -15,21 +15,33 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const commentModel_1 = __importDefault(require("../model/commentModel"));
 const profileModel_1 = __importDefault(require("../model/profileModel"));
 const userModel_1 = __importDefault(require("../model/userModel"));
+const mongoose_1 = __importDefault(require("mongoose"));
 class CommentService {
     // 댓글 저장 서비스
     saveComment(user_id, data) {
         return __awaiter(this, void 0, void 0, function* () {
+            if (!mongoose_1.default.Types.ObjectId.isValid(user_id)) {
+                console.error("유효하지 않은 user_id입니다.");
+                return null;
+            }
             const body = Object.assign(Object.assign({}, data), { user_id: user_id });
-            const comment = new commentModel_1.default(body);
-            if (comment) {
+            try {
+                const comment = new commentModel_1.default(body);
                 return yield comment.save();
             }
-            return null;
+            catch (error) {
+                console.error("댓글 저장 에러:", error);
+                return null;
+            }
         });
     }
     // 댓글 조회 서비스
     getComment(post_id) {
         return __awaiter(this, void 0, void 0, function* () {
+            if (!mongoose_1.default.Types.ObjectId.isValid(post_id)) {
+                console.error("유효하지 않은 post_id입니다.");
+                return null;
+            }
             try {
                 const comments = yield commentModel_1.default.find({ post_id }).exec();
                 if (!comments)
@@ -57,6 +69,11 @@ class CommentService {
     // 댓글 수정 서비스
     updateComment(comment_id, user_id, data) {
         return __awaiter(this, void 0, void 0, function* () {
+            if (!mongoose_1.default.Types.ObjectId.isValid(comment_id) ||
+                !mongoose_1.default.Types.ObjectId.isValid(user_id)) {
+                console.error("유효하지 않은 comment_id 또는 user_id입니다.");
+                return null;
+            }
             try {
                 const updateData = Object.assign(Object.assign({}, data), { updated_date: new Date() });
                 const update = yield commentModel_1.default.findOneAndUpdate({ comment_id: comment_id, user_id: user_id }, { $set: updateData }, { new: true }).exec();
@@ -74,6 +91,11 @@ class CommentService {
     // 댓글 삭제
     deleteComment(comment_id, user_id) {
         return __awaiter(this, void 0, void 0, function* () {
+            if (!mongoose_1.default.Types.ObjectId.isValid(comment_id) ||
+                !mongoose_1.default.Types.ObjectId.isValid(user_id)) {
+                console.error("유효하지 않은 comment_id 또는 user_id입니다.");
+                return null;
+            }
             const deleteComment = yield commentModel_1.default.findOneAndDelete({
                 comment_id,
                 user_id,
